Pass user-specified parameters to the Linux kernel. Any options to be passed to the kernel that have been specified in the .MLUL-Live-USB are passed.

+++ src/main.h
@@ -45,7 +45,7 @@ typedef struct BootableLinuxDistro {
 	struct BootableLinuxDistro *next;
 } BootableLinuxDistro;
 
-EFI_STATUS BootLinuxWithOptions(CHAR16 *, int);
+EFI_STATUS BootLinuxWithOptions(CHAR16 *, UINT16);
 
 extern const EFI_GUID enterprise_variable_guid;
 extern const EFI_GUID grub_variable_guid;
